app/requirements.txt
flask
simplepyble
pygame
pygobject
RPi.GPIO
gunicorn